raspberrypi/camera/requirements.txt
opencv-python
numpy
tflite-runtime; platform_machine == 'armv7l' or platform_machine == 'aarch64'
tflite-runtime==2.5.0; platform_system == 'Windows'
pyserial
mediapipe
